+++ app/src/main/java/com/example/gridgambit/GridUtil.java
@@ -9,6 +9,7 @@ import android.util.Log;
 import android.util.TypedValue;
 import android.view.Gravity;
 import android.view.View;
+import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.LinearLayout;
 import android.widget.TextView;
@@ -84,7 +85,156 @@ public class GridUtil {
         return gridLock;
     }
 
-    // TODO: Add wincheck
+    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    public static void winCheck(final Context context, final Level levelUI, final Activity activity){
+
+        if (Level.LevelInfo.currentScore >= Level.LevelInfo.targetScore && (Player.PlayerInfo.level != 14) && (Player.PlayerInfo.level != 17) || (Player.PlayerInfo.level == 14 && Level.LevelInfo.currentScore <= Level.LevelInfo.targetScore) || (Player.PlayerInfo.level == 17 && Level.LevelInfo.currentScore <= Level.LevelInfo.targetScore)){
+            //TODO: add sound
+            final LinearLayout winScreen = new LinearLayout(context);
+            Button mainMenuButton = new Button(context);
+            mainMenuButton.setText(R.string.main_menu);
+            mainMenuButton.setTextSize(TypedValue.COMPLEX_UNIT_SP, 25);
+            mainMenuButton.setTextColor(Color.parseColor("#dddddd"));
+            mainMenuButton.setBackground(context.getDrawable(R.drawable.endgame_button_background));
+            Button closeButton = new Button(context);
+            closeButton.setText("");
+            closeButton.setTextSize(TypedValue.COMPLEX_UNIT_SP, 25);
+            closeButton.setGravity(Gravity.CENTER);
+            closeButton.setTextColor(Color.parseColor("#dddddd"));
+            closeButton.setBackground(context.getDrawable(R.drawable.close_drawable));
+            TextView winMessage = new TextView(context);
+            winMessage.setTextSize(TypedValue.COMPLEX_UNIT_SP, 25);
+            winMessage.setTextColor(Color.parseColor("#dddddd"));
+            winMessage.setGravity(Gravity.CENTER);
+            final Button nextLevelButton = new Button(context);
+            nextLevelButton.setText(R.string.next_level);
+            nextLevelButton.setTextSize(TypedValue.COMPLEX_UNIT_SP, 25);
+            nextLevelButton.setTextColor(Color.parseColor("#dddddd"));
+            nextLevelButton.setBackground(context.getDrawable(R.drawable.endgame_button_background));
+
+            mainMenuButton.setOnClickListener(new View.OnClickListener() {
+
+                public void onClick(View v) {
+                    Level.LevelInfo.soundPool.release();
+                    Level.LevelInfo.soundPool = null;
+
+                    Intent intent = new Intent(v.getContext(), MainActivity.class);
+                    context.startActivity(intent);
+                    activity.finish();
+
+                }
+            });
+
+            closeButton.setOnClickListener(new View.OnClickListener() {
+
+                public void onClick(View v) {
+                    Player.PlayerInfo.levelPassed = true;
+                    Level.LevelInfo.targetScore = 99999;
+                    Level.LevelInfo.currentTurns = 99999;
+                    LinearLayout row1 = (LinearLayout) levelUI.turnsText.getParent();
+                    LinearLayout row2 = (LinearLayout) levelUI.scoreText.getParent();
+                    row1.removeView(levelUI.turnsText);
+                    row1.removeView(levelUI.levelText);
+                    row2.removeView(levelUI.targetText);
+
+                    ((ViewGroup) winScreen.getParent()).removeView(winScreen);
+
+                    LinearLayout nextLevelParent = (LinearLayout) nextLevelButton.getParent();
+                    if(Player.PlayerInfo.level < Level.LevelInfo.levelArray.length()-1) {
+                        nextLevelParent.removeView(nextLevelButton);
+                        float density = v.getContext().getResources().getDisplayMetrics().density;
+                        int layoutMargin = (int) (10 * density + 0.5f);
+                        LinearLayout.LayoutParams params = new LinearLayout.LayoutParams(ConstraintLayout.LayoutParams.MATCH_PARENT, ConstraintLayout.LayoutParams.WRAP_CONTENT);
+                        params.bottomMargin = layoutMargin;
+                        nextLevelButton.setLayoutParams(params);
+                        nextLevelButton.setTextSize(TypedValue.COMPLEX_UNIT_SP, 20);
+                        row1.addView(nextLevelButton);
+                    }
+                }
+            });
+
+            nextLevelButton.setOnClickListener(new View.OnClickListener() {
+
+                public void onClick(View v) {
+                    Level.LevelInfo.soundPool.release();
+                    Level.LevelInfo.soundPool = null;
+                    // TODO: add save player data
+                    Player.PlayerInfo.levelPassed = false;
+                    Intent intent = new Intent(v.getContext(), GameScreen.class);
+                    intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                    context.startActivity(intent);
+                    activity.finish();
+                }
+            });
+
+            if(Player.PlayerInfo.level < Level.LevelInfo.levelArray.length()-1){
+
+                if(!Player.PlayerInfo.isEndless){
+
+                    if(Player.PlayerInfo.levelHighest < Player.PlayerInfo.level){
+
+                        Player.PlayerInfo.levelHighest = Player.PlayerInfo.level;
+                    }
+                }
+                else {
+                    Player.PlayerInfo.level = 0;
+                }
+
+                winMessage.setText(String.format(Locale.getDefault(), "%s %d %s", context.getString(R.string.level_text), (Player.PlayerInfo.level+1), " complete!"));
+            }
+            else{
+                winMessage.setText(R.string.finish);
+            }
+
+            if(Player.PlayerInfo.level < Level.LevelInfo.levelArray.length()-1){
+                winScreen.setOrientation(LinearLayout.VERTICAL);
+                winScreen.addView(closeButton);
+                winScreen.addView(winMessage);
+                winScreen.addView(nextLevelButton);
+                winScreen.addView(mainMenuButton);
+                Player.PlayerInfo.level++;
+            }
+            else{
+                winScreen.setOrientation(LinearLayout.VERTICAL);
+                winScreen.addView(closeButton);
+                winScreen.addView(winMessage);
+                winScreen.addView(mainMenuButton);
+            }
+
+            winScreen.setBackground(context.getDrawable(R.drawable.status_text_background));
+            winScreen.setClickable(true);
+            winScreen.setId(View.generateViewId());
+            ConstraintLayout constraintLayout = activity.findViewById(R.id.game_constraint_layout);
+            constraintLayout.addView(winScreen);
+            winScreen.setElevation(100);
+
+            float density = context.getResources().getDisplayMetrics().density;
+            int layoutMargin = (int) (50 * density + 0.5f);
+            ConstraintSet cs = new ConstraintSet();
+            cs.clone(constraintLayout);
+            cs.connect(winScreen.getId(), ConstraintSet.TOP, R.id.info_ui_layout, ConstraintSet.TOP, 0);
+            cs.applyTo(constraintLayout);
+            ConstraintLayout.LayoutParams params = new ConstraintLayout.LayoutParams(ConstraintLayout.LayoutParams.MATCH_PARENT, ConstraintLayout.LayoutParams.MATCH_PARENT);
+            LinearLayout.LayoutParams paramsCloseButton = new LinearLayout.LayoutParams(layoutMargin, layoutMargin);
+            LinearLayout.LayoutParams paramsNextLevelButton = new LinearLayout.LayoutParams(ConstraintLayout.LayoutParams.MATCH_PARENT, ConstraintLayout.LayoutParams.WRAP_CONTENT);
+            LinearLayout.LayoutParams paramsMainMenuButton = new LinearLayout.LayoutParams(ConstraintLayout.LayoutParams.MATCH_PARENT, ConstraintLayout.LayoutParams.WRAP_CONTENT);
+            LinearLayout.LayoutParams paramsWinMessage = new LinearLayout.LayoutParams(ConstraintLayout.LayoutParams.MATCH_PARENT, ConstraintLayout.LayoutParams.WRAP_CONTENT);
+            paramsWinMessage.bottomMargin = layoutMargin;
+            paramsNextLevelButton.bottomMargin = layoutMargin;
+            mainMenuButton.setLayoutParams(paramsMainMenuButton);
+            winMessage.setLayoutParams(paramsWinMessage);
+            nextLevelButton.setLayoutParams(paramsNextLevelButton);
+
+            winScreen.setLayoutParams(params);
+            winScreen.setGravity(Gravity.END | Gravity.CENTER);
+            closeButton.setLayoutParams(paramsCloseButton);
+
+            // TODO: add achievement check
+        }
+        else{
+            loseCheck(context, levelUI, activity);
+        }
+    }
 
     @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
     public static void loseCheck(final Context context, Level levelUI, final Activity activity){